+++ 04-engine-op/src/main/scala/orchescala/engine/op/OperatonHistoricProcessInstanceService.scala
@@ -1,13 +1,13 @@
-package orchescala.engine.operaton
+package orchescala.engine.op
 
 import orchescala.engine.EngineConfig
-import orchescala.engine.c7.C7ProcessInstanceService
+import orchescala.engine.c7.C7HistoricProcessInstanceService
 import orchescala.engine.domain.EngineError
 import org.camunda.community.rest.client.invoker.ApiClient
 import zio.IO
 
-class OperatonProcessInstanceService(using
+class OperatonHistoricProcessInstanceService(using
     apiClientZIO: IO[EngineError, ApiClient],
     engineConfig: EngineConfig
-) extends C7ProcessInstanceService, OperatonService, OperatonEventService
+) extends C7HistoricProcessInstanceService, OperatonService
