+++ 04-engine-op/src/main/scala/orchescala/engine/op/OperatonMessageService.scala
@@ -1,13 +1,13 @@
-package orchescala.engine.operaton
+package orchescala.engine.op
 
 import orchescala.engine.EngineConfig
-import orchescala.engine.c7.C7JobService
+import orchescala.engine.c7.C7MessageService
 import orchescala.engine.domain.EngineError
 import org.camunda.community.rest.client.invoker.ApiClient
 import zio.IO
 
-class OperatonJobService(using
+class OperatonMessageService(using
     apiClientZIO: IO[EngineError, ApiClient],
     engineConfig: EngineConfig
-) extends C7JobService, OperatonService
+) extends C7MessageService, OperatonService, OperatonEventService
