+++ 04-engine-op/src/main/scala/orchescala/engine/op/OperatonService.scala
@@ -1,4 +1,4 @@
-package orchescala.engine.operaton
+package orchescala.engine.op
 
 import orchescala.engine.c7.C7Service
 import orchescala.engine.domain.EngineType
