+++ 04-engine-op/src/main/scala/orchescala/engine/op/OperatonSignalService.scala
@@ -1,4 +1,4 @@
-package orchescala.engine.operaton
+package orchescala.engine.op
 
 import orchescala.engine.EngineConfig
 import orchescala.engine.c7.C7SignalService
