+++ 04-engine-op/src/main/scala/orchescala/engine/op/SharedOperatonClientManager.scala
@@ -1,4 +1,4 @@
-package orchescala.engine.operaton
+package orchescala.engine.op
 
 import orchescala.engine.SharedClientManager
 import orchescala.engine.domain.EngineError
@@ -6,19 +6,19 @@ import org.camunda.community.rest.client.invoker.ApiClient
 import zio.*
 
 /** Service trait for managing shared Operaton API Client for simulations */
-type SharedOperatonClientManager = SharedClientManager[ApiClient, EngineError]
+type SharedOpClientManager = SharedClientManager[ApiClient, EngineError]
 
 object SharedOperatonClientManager:
 
   /** ZLayer that provides SharedOperatonClientManager service */
-  val layer: ZLayer[Any, Nothing, SharedOperatonClientManager] =
+  val layer: ZLayer[Any, Nothing, SharedOpClientManager] =
     SharedClientManager.createLayer[ApiClient, EngineError](
       "Operaton API",
       client => ZIO.attempt(client.getHttpClient.close()).ignore
     )
 
   /** Convenience method to access the service */
-  def getOrCreateClient(clientFactory: ZIO[Any, EngineError, ApiClient]): ZIO[SharedOperatonClientManager, EngineError, ApiClient] =
+  def getOrCreateClient(clientFactory: ZIO[Any, EngineError, ApiClient]): ZIO[SharedOpClientManager, EngineError, ApiClient] =
     SharedClientManager.getOrCreateClient(clientFactory)
 
 end SharedOperatonClientManager
