+++ 04-worker-op/src/main/scala/orchescala/worker/op/OperatonWorkerClient.scala
@@ -1,4 +1,4 @@
-package orchescala.worker.operaton
+package orchescala.worker.op
 
 import orchescala.domain.OrchescalaLogger
 import orchescala.engine.Slf4JLogger
